requirements.txt
# Core framework
langchain-core==0.3.63
langgraph==0.5.4
langchain==0.3.25
langchain-community==0.3.24

# Model providers
langchain-openai==0.3.19
langchain-anthropic==0.3.15
langchain-google-genai==2.1.5

# Tools
tavily-python==0.5.0

# Web framework
streamlit==1.40.1

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3

# Monitoring
langsmith==0.1.147
